# openapi/local-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CSGHub Lite Local API",
    "version": "0.1.0",
    "description": "Structured documentation for the local `/api/*` and `/v1/*` endpoints served by `csghub-lite`.\n\nMaintenance note: whenever you add, remove, or rename an `/api/*` or `/v1/*` route in `internal/server/routes.go`, update this file and run the OpenAPI sync test before merging."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current csghub-lite server"
    },
    {
      "url": "http://localhost:11435",
      "description": "Default local server"
    }
  ],
...
  "paths": {
...
    "/api/apps/provider": {
      "post": {
        "tags": [
          "AI Apps"
        ],
        "summary": "Switch an AI App model provider",
        "description": "Switches Codex or Claude Code between its native provider configuration and OpenCSG. Codex CLI and Codex App share one provider state.",
        "operationId": "switchAIAppProvider",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIAppActionRequest"
              },
              "example": {
                "app_id": "codex-app",
                "provider_mode": "opencsg",
                "model_id": "Qwen/Qwen3-0.6B-GGUF"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider switched; refreshed app state returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIAppInfo"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "description": "The app configuration changed after OpenCSG was enabled and was not overwritten."
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "The request body, query parameters, or path parameters were invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error",
          "errorCode"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "errorCode": {
            "type": "integer",
            "format": "int32",
            "example": 400
          }
        }
      },
...
      "AIAppActionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "app_id"
        ],
        "properties": {
          "app_id": {
            "type": "string"
          },
          "model_id": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "description": "Model source such as local, cloud, or provider:<id>."
          },
          "provider_mode": {
            "type": "string",
            "enum": [
              "native",
              "opencsg"
            ],
            "description": "Provider mode requested for apps that support explicit source switching."
          },
          "model_bindings": {
            "type": "array",
            "description": "Task-specific model selections used by multi-model apps such as Xiaozhi.",
            "items": {
              "$ref": "#/components/schemas/AIAppModelBinding"
            }
          },
          "work_dir": {
            "type": "string",
            "description": "Working directory used for shell-based AI App launches."
          }
        }
      },
      "AIAppInfo": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "installed",
          "managed",
          "supported",
          "disabled",
          "status",
          "progress_mode",
          "runtime_supported",
          "runtime_running",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "installed": {
            "type": "boolean"
          },
          "managed": {
            "type": "boolean"
          },
          "supported": {
            "type": "boolean"
          },
          "disabled": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "idle",
              "installing",
              "uninstalling",
              "installed",
              "failed",
              "disabled"
            ]
          },
          "phase": {
            "type": "string"
          },
          "progress_mode": {
            "type": "string",
            "enum": [
              "percent",
              "indeterminate"
            ]
          },
          "progress": {
            "type": "integer"
          },
          "install_path": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "latest_version": {
            "type": "string"
          },
          "update_available": {
            "type": "boolean"
          },
          "model_id": {
            "type": "string"
          },
          "model_source": {
            "type": "string",
            "description": "Persisted model source such as local, cloud, or provider:<id>."
          },
          "provider_mode": {
            "type": "string",
            "enum": [
              "native",
              "opencsg"
            ]
          },
          "provider_group": {
            "type": "string",
            "description": "Shared provider state group. Codex CLI and Codex App both use codex."
          },
          "provider_switch_supported": {
            "type": "boolean"
          },
          "provider_drifted": {
            "type": "boolean",
            "description": "Whether the live app configuration changed after OpenCSG was enabled."
          },
          "model_bindings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AIAppModelBinding"
            }
          },
          "model_slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AIAppModelSlot"
            }
          },
          "runtime_supported": {
            "type": "boolean",
            "description": "Whether this app exposes a long-running runtime that can be started or stopped."
          },
          "runtime_running": {
            "type": "boolean",
            "description": "Whether the app runtime is currently reachable."
          },
          "runtime_status": {
            "type": "string",
            "enum": [
              "running",
              "stopped"
            ]
          },
          "log_path": {
            "type": "string"
          },
          "last_error": {
            "type": "string"
          },
          "disabled_reason": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AIAppModelBinding": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "task",
          "model_id",
          "source"
        ],
        "properties": {
          "task": {
            "type": "string",
            "enum": [
              "language_model",
              "speech_recognition",
              "embedding",
              "image_generation"
            ]
          },
          "model_id": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "description": "Model source such as local, cloud, or provider:<id>."
          }
        }
      },
      "AIAppModelSlot": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "task",
          "required"
        ],
        "properties": {
          "task": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "binding": {
            "$ref": "#/components/schemas/AIAppModelBinding"
          }
        }
      },
...
    }
  }
}